Added detailed weather interface, create a separate FetchWeatherTask class
