+++ app/src/main/res/layout/fragment_detail.xml
@@ -8,11 +8,118 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="com.example.vs9x64bit.weatherapplication.DetailActivityFragment"
     tools:showIn="@layout/activity_detail">
+    <android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:card_view="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/card_view"
+        android:layout_width="match_parent"
+        android:layout_height="230dp"
+        android:layout_marginLeft="0dp"
+        android:layout_marginRight="0dp"
+        android:layout_marginTop="10dp"
+        android:layout_marginBottom="10dp"
+        card_view:cardCornerRadius="4dp"
+        android:elevation="5dp"
+        android:clickable="true"
+        android:foreground="?android:attr/selectableItemBackground">
 
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <TextView
+                android:id="@+id/detailed_location"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="location"
+                android:textSize="22sp"
+                android:layout_alignParentTop="true"
+                android:layout_centerHorizontal="true"
+                android:layout_marginTop="15dp" />
+
+            <TextView
+                android:id="@+id/detailed_description"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="description"
+                android:gravity="center"
+                android:textSize="12sp"
+                android:layout_below="@+id/detailed_location"
+                android:layout_alignLeft="@+id/detailed_location"
+                android:layout_alignRight="@+id/detailed_location"
+             />
+
+
+            <TextView
+                android:id="@+id/detailed_temperature"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="20"
+                android:gravity="center"
+                android:textSize="60sp"
+                android:layout_alignLeft="@+id/detailed_description"
+                android:layout_alignRight="@+id/detailed_description"
+                android:layout_below="@+id/detailed_description" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Pressure: "
+                android:id="@+id/detailed_pressure"
+                android:textSize="12sp"
+                android:layout_below="@+id/detailed_temperature"
+                android:layout_alignLeft="@+id/detailed_humidity"
+                android:layout_alignStart="@+id/detailed_humidity" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Humidity: "
+                android:id="@+id/detailed_humidity"
+                android:textSize="12sp"
+                android:layout_marginBottom="23dp"
+                android:layout_alignParentBottom="true"
+                android:layout_alignParentLeft="true"
+                android:layout_alignParentStart="true"
+                android:layout_marginLeft="25dp"
+                android:layout_marginStart="25dp" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Precipitation: "
+                android:textSize="12sp"
+                android:id="@+id/detailed_precipitation"
+                android:layout_alignTop="@+id/detailed_pressure"
+                android:layout_toRightOf="@+id/detailed_location"
+                android:layout_toEndOf="@+id/detailed_location" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Wind: "
+                android:textSize="12sp"
+                android:id="@+id/detailed_wind"
+                android:layout_alignTop="@+id/detailed_humidity"
+                android:layout_alignLeft="@+id/detailed_precipitation"
+                android:layout_alignStart="@+id/detailed_precipitation" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="°"
+                android:textSize="25sp"
+                android:id="@+id/detailed_sign"
+                android:layout_alignTop="@+id/detailed_temperature"
+                android:layout_alignRight="@+id/detailed_temperature"
+                android:layout_alignEnd="@+id/detailed_temperature" />
+
+        </RelativeLayout>
+
+    </android.support.v7.widget.CardView>
     <ListView
         android:id="@+id/listview_forecast"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="228dp"
         android:layout_gravity="right|bottom">
     </ListView>
